Extract channel error handlers
Refactor the channel state machine to extract the error handlers and a few
utility handlers to separate files.

The functions moved are completely unchanged, except for their visibility
(going from private to public).
